+++ Modules/Sources/JetpackStats/Services/CSVExporter.swift
@@ -1,4 +1,5 @@
 import Foundation
+import SwiftUI
 
 protocol CSVExporterProtocol {
     func generateCSV(from items: [any TopListItemProtocol], metric: SiteMetric) -> String
@@ -64,5 +65,31 @@ struct CSVExporter: CSVExporterProtocol {
         // Escape quotes by doubling them and wrap the field in quotes
         let escaped = field.replacingOccurrences(of: "\"", with: "\"\"")
         return "\"\(escaped)\""
+    }
+}
+
+struct CSVDataRepresentation: Transferable {
+    let items: [any TopListItemProtocol]
+    let metric: SiteMetric
+    let fileName: String
+
+    static var transferRepresentation: some TransferRepresentation {
+        let dataRepresentation = DataRepresentation(exportedContentType: .commaSeparatedText) { (representation: CSVDataRepresentation) in
+            try representation.generateCSVData()
+        }
+        if #available(iOS 17.0, *) {
+            return dataRepresentation.suggestedFileName { $0.fileName }
+        } else {
+            return dataRepresentation
+        }
+    }
+
+    private func generateCSVData() throws -> Data {
+        let exporter = CSVExporter()
+        let csvContent = exporter.generateCSV(from: items, metric: metric)
+        guard let data = csvContent.data(using: .utf8) else {
+            throw CocoaError(.fileWriteUnknown)
+        }
+        return data
     }
 }
